Refactor codebase: extract shared CSS, split reducer, clean up dead code
- Delete dead GameHistory component and unused react.svg asset
- Extract duplicated CSS into shared.css (scoreboard, bust colors, badges, banners)
- Create BustChanceBanner component replacing duplicate IIFE blocks
- Extract flattenHandWithCancelled utility into handUtils.js
- Add createEmptyDealtCards to deckUtils.js
- Split play-mode actions into separate playModeReducer.js (~450 lines)
- Remove global SET_DIALOG in favor of local useState in components
- Standardize NAVIGATE payload to { screen, editingRound?, playerId? }
- Remove unused getDealtCards/getDeckStatus helpers
- Move test-helpers.js to test-utils/helpers.js

Fixes FLI-9
