+++ src/components/GameScreen/GameScreen.module.css
@@ -1,108 +1,1 @@
-.gameHeader {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 20px 0 12px 0;
-  position: sticky;
-  top: 0;
-  background: var(--bg);
-  z-index: 10;
-}
-
-.gameTitle {
-  font-family: 'Fredoka', sans-serif;
-  font-size: 24px;
-  font-weight: 700;
-  color: var(--primary);
-}
-
-.roundBadge {
-  background: var(--surface2);
-  padding: 6px 14px;
-  border-radius: 20px;
-  font-size: 13px;
-  font-weight: 600;
-  color: var(--text-dim);
-}
-
-.winnerBanner {
-  background: linear-gradient(135deg, #ffb347, #ff6b6b);
-  border-radius: var(--radius);
-  padding: 20px;
-  text-align: center;
-  margin-bottom: 16px;
-  animation: pop 0.4s ease-out;
-}
-
-.winnerBanner h3 {
-  font-size: 24px;
-  color: #fff;
-}
-
-.winnerBanner p {
-  color: rgba(255,255,255,0.85);
-  font-size: 14px;
-}
-
-.scoreboard { margin-bottom: 16px; }
-
-.playerRow {
-  display: flex;
-  align-items: center;
-  gap: 12px;
-  padding: 14px 16px;
-  background: var(--surface);
-  border-radius: var(--radius-sm);
-  margin-bottom: 8px;
-  cursor: pointer;
-  transition: all 0.2s;
-  border: 2px solid transparent;
-}
-
-.playerRow:hover { border-color: var(--primary); }
-.playerRow.leader { border-color: var(--accent); }
-
-.rank {
-  font-family: 'Fredoka', sans-serif;
-  font-weight: 700;
-  font-size: 18px;
-  color: var(--text-dim);
-  width: 28px;
-  text-align: center;
-}
-
-.rank.gold { color: var(--accent); }
-
-.playerInfo { flex: 1; }
-
-.playerName {
-  font-weight: 600;
-  font-size: 16px;
-}
-
-.playerLastRound {
-  font-size: 12px;
-  color: var(--text-dim);
-  margin-top: 2px;
-}
-
-.playerScore {
-  font-family: 'Fredoka', sans-serif;
-  font-weight: 700;
-  font-size: 28px;
-  color: var(--primary);
-}
-
-.playerRow.leader .playerScore { color: var(--accent); }
-
-.arrowIcon {
-  color: var(--text-dim);
-  font-size: 18px;
-}
-
-.actionBar {
-  display: flex;
-  gap: 10px;
-  margin: 16px 0;
-  flex-wrap: wrap;
-}
+/* GameScreen-specific overrides — shared scoreboard styles are in shared.css */
