+++ src/components/RoundEntry/RoundEntry.module.css
@@ -12,27 +12,6 @@
   margin: 0;
 }
 
-.cheaterToggle {
-  font-family: 'Fredoka', sans-serif;
-  font-size: 11px;
-  font-weight: 600;
-  padding: 6px 12px;
-  border-radius: 20px;
-  border: 1px solid var(--surface2);
-  background: var(--surface2);
-  color: var(--text-dim);
-  cursor: pointer;
-  transition: all 0.2s;
-  text-transform: uppercase;
-  letter-spacing: 0.5px;
-}
-
-.cheaterActive {
-  border-color: var(--accent);
-  background: rgba(255, 179, 71, 0.15);
-  color: var(--accent);
-}
-
 .playerSection {
   background: var(--surface);
   border-radius: var(--radius);
@@ -209,51 +188,6 @@
   opacity: 0.4;
   cursor: not-allowed;
   pointer-events: none;
-}
-
-.bustChanceBanner {
-  display: flex;
-  align-items: center;
-  gap: 10px;
-  padding: 10px 14px;
-  border-radius: var(--radius-sm);
-  margin-bottom: 12px;
-  border: 1px solid;
-}
-
-.bustChanceValue {
-  font-family: 'Fredoka', sans-serif;
-  font-weight: 700;
-  font-size: 20px;
-}
-
-.bustChanceLabel {
-  font-size: 12px;
-  opacity: 0.85;
-}
-
-.bustSafe {
-  background: rgba(0, 212, 170, 0.1);
-  border-color: rgba(0, 212, 170, 0.3);
-  color: var(--primary);
-}
-
-.bustCaution {
-  background: rgba(255, 179, 71, 0.1);
-  border-color: rgba(255, 179, 71, 0.3);
-  color: var(--accent);
-}
-
-.bustWarning {
-  background: rgba(255, 152, 71, 0.1);
-  border-color: rgba(255, 152, 71, 0.3);
-  color: #ff9847;
-}
-
-.bustDanger {
-  background: rgba(255, 107, 107, 0.1);
-  border-color: rgba(255, 107, 107, 0.3);
-  color: var(--accent2);
 }
 
 .dealTimeline {
