+++ src/components/SetupScreen/SetupScreen.module.css
@@ -110,19 +110,6 @@
   border-color: rgba(255, 179, 71, 0.3);
 }
 
-.cpuBadge {
-  font-size: 10px;
-  font-weight: 700;
-  padding: 2px 8px;
-  border-radius: 10px;
-  background: rgba(255, 179, 71, 0.15);
-  color: var(--accent);
-  font-family: 'Fredoka', sans-serif;
-  text-transform: uppercase;
-  letter-spacing: 0.5px;
-  flex-shrink: 0;
-}
-
 .addCpuBtn {
   border-color: var(--accent) !important;
   color: var(--accent) !important;
