+++ src/styles/global.css
@@ -1,4 +1,5 @@
 @import url('https://fonts.googleapis.com/css2?family=Fredoka:wght@400;500;600;700&family=DM+Sans:wght@400;500;600;700&display=swap');
+@import './shared.css';
 
 :root {
   --bg: #0f1923;
